+++ src/main/kotlin/care/better/tools/aqlidea/plugin/toolWindow/AqlToolWindowFactory.kt
@@ -1,6 +1,7 @@
 package care.better.tools.aqlidea.plugin.toolWindow
 
 import care.better.tools.aqlidea.plugin.toolWindow.query.AqlQueryResultHeaderBuilder
+import care.better.tools.aqlidea.plugin.toolWindow.query.AqlQueryPanel
 import care.better.tools.aqlidea.plugin.toolWindow.servers.AqlServersPanel
 import care.better.tools.aqlidea.thinkehr.ThinkEhrClient
 import care.better.tools.aqlidea.ui.treetable.TreeTableData
@@ -15,7 +16,7 @@ class AqlToolWindowFactory: ToolWindowFactory {
 
 
     override fun createToolWindowContent(project: Project, toolWindow: ToolWindow) {
-        val aqlToolWindow = AqlToolWindow()
+        val aqlToolWindow = AqlQueryPanel(project)
 
         val contentFactory = ContentFactory.SERVICE.getInstance()
         val serversPanel = AqlServersPanel(project)
@@ -38,43 +39,44 @@ class AqlToolWindowFactory: ToolWindowFactory {
 
         fun showQueryStart(project: Project) {
             val toolWindow = getToolWindow(project)
-            val aqlToolWindow = toolWindow.contentManager.contents[1].component as AqlToolWindow
+            val aqlToolWindow = toolWindow.contentManager.contents[1].component as AqlQueryPanel
             toolWindow.contentManager.setSelectedContent(toolWindow.contentManager.contents[1])
-            aqlToolWindow.showQueryTab(AqlToolWindow.QueryTab.table)
-            aqlToolWindow.setRawRequest(null)
-            aqlToolWindow.setRawResponse("<pending>")
+            aqlToolWindow.showQueryTab(AqlQueryPanel.QueryTab.table)
+            aqlToolWindow.rawRequest = ""
+            aqlToolWindow.rawResponse = "<pending>"
             aqlToolWindow.setResultTableData(null)
             toolWindow.activate { }
         }
 
         fun showQueryError(project: Project, e: Exception) {
             val toolWindow = getToolWindow(project)
-            val aqlToolWindow = toolWindow.contentManager.contents[1].component as AqlToolWindow
+            val aqlToolWindow = toolWindow.contentManager.contents[1].component as AqlQueryPanel
             toolWindow.contentManager.setSelectedContent(toolWindow.contentManager.contents[1])
             toolWindow.activate {
                 if (e is ThinkEhrClient.ThinkEhrAqlException) {
-                    aqlToolWindow.setRawRequest(e.request.body)
-                    aqlToolWindow.setRawResponse(e.response?.body)
+                    aqlToolWindow.rawRequest = e.request.body ?: ""
+                    aqlToolWindow.rawResponse = e.response?.body ?: ""
                     aqlToolWindow.setResultTableData(null)
-                    aqlToolWindow.showQueryTab(AqlToolWindow.QueryTab.response)
+                    aqlToolWindow.showQueryTab(AqlQueryPanel.QueryTab.response)
                 }
                 else {
                     log.error("Error calling ThinkEhr server", e)
-                    aqlToolWindow.setRawRequest("")
-                    aqlToolWindow.setRawResponse(e.toString())
+                    aqlToolWindow.rawRequest = ""
+                    aqlToolWindow. rawResponse = e.toString()
                     aqlToolWindow.setResultTableData(null)
-                    aqlToolWindow.showQueryTab(AqlToolWindow.QueryTab.response)
+                    aqlToolWindow.showQueryTab(AqlQueryPanel.QueryTab.response)
                 }
             }
         }
 
         fun showQueryResult(project: Project, response: ThinkEhrClient.QueryResponse) {
             val toolWindow = getToolWindow(project)
-            val aqlToolWindow = toolWindow.contentManager.contents[1].component as AqlToolWindow
+            val aqlToolWindow = toolWindow.contentManager.contents[1].component as AqlQueryPanel
             toolWindow.contentManager.setSelectedContent(toolWindow.contentManager.contents[1])
             toolWindow.activate {
-                aqlToolWindow.setRawRequest(response.rawRequest.body)
-                aqlToolWindow.setRawResponse(response.rawResponse.body)
+                aqlToolWindow.rawRequest = response.rawRequest.body ?: ""
+                aqlToolWindow.rawResponse = response.rawResponse.body ?: ""
+
                 updateTableValues(aqlToolWindow, response)
             }
         }
@@ -82,17 +84,17 @@ class AqlToolWindowFactory: ToolWindowFactory {
         private fun getToolWindow(project: Project) =
             ToolWindowManager.getInstance(project).getToolWindow(TOOL_WINDOW_ID)!!
 
-        private fun updateTableValues(aqlToolWindow: AqlToolWindow, response: ThinkEhrClient.QueryResponse) {
+        private fun updateTableValues(aqlToolWindow: AqlQueryPanel, response: ThinkEhrClient.QueryResponse) {
             val parsedResponse = response.response
 
             if (parsedResponse != null) {
                 val treeTableData = AqlQueryResultHeaderBuilder().build(parsedResponse)
                 aqlToolWindow.setResultTableData(treeTableData)
 
-                aqlToolWindow.showQueryTab(AqlToolWindow.QueryTab.table)
+                aqlToolWindow.showQueryTab(AqlQueryPanel.QueryTab.table)
             } else {
                 aqlToolWindow.setResultTableData(TreeTableData(listOf(), listOf()))
-                aqlToolWindow.showQueryTab(AqlToolWindow.QueryTab.response)
+                aqlToolWindow.showQueryTab(AqlQueryPanel.QueryTab.response)
 
             }
         }